pyproject.toml
[project]
name = "code-doc-assistant"
version = "0.1.0"
description = "Conversational AI assistant that ingests a codebase and answers questions about it using RAG"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.2",
    "fastapi>=0.135.1",
    "gitpython>=3.1.46",
    "httpx>=0.28.1",
    "openai>=2.24.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "rich>=14.3.3",
    "streamlit>=1.55.0",
    "tiktoken>=0.12.0",
    "uvicorn[standard]>=0.41.0",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/code_doc_assistant"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--cov=src/code_doc_assistant --cov-report=term-missing --cov-report=html"

[tool.coverage.run]
source = ["src/code_doc_assistant"]
omit = ["tests/*"]
